+++ src/extension.ts
@@ -95,6 +95,14 @@ export function activate(_context: vscode.ExtensionContext) {
 		await dmc.createEmptyObject(bucket, context);
 		simpleStorageDataProvider.refresh();
 	});
+	vscode.commands.registerCommand('forge.copyObject', async (object?: IObject) => {
+		await dmc.copyObject(object, context);
+		simpleStorageDataProvider.refresh();
+	});
+	vscode.commands.registerCommand('forge.renameObject', async (object?: IObject) => {
+		await dmc.renameObject(object, context);
+		simpleStorageDataProvider.refresh();
+	});
 	vscode.commands.registerCommand('forge.downloadObject', async (object?: IObject) => {
 		await dmc.downloadObject(object, context);
 	});